Create app component and refactor index.jsx
